Struct google_manufacturers1::AccountMethods

pub struct AccountMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on account resources. It is not used directly, but through the ManufacturerCenter hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_manufacturers1 as manufacturers1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use manufacturers1::ManufacturerCenter;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = ManufacturerCenter::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `products_delete(...)`, `products_get(...)`, `products_list(...)` and `products_update(...)`
// to build up your call.
let rb = hub.accounts();

Methods

impl<'a, C, A> AccountMethods<'a, C, A>

pub fn products_list(&self, parent: &str) -> AccountProductListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all the products in a Manufacturer Center account.

Arguments

parent - Parent ID in the format accounts/{account_id}. account_id - The ID of the Manufacturer Center account.

pub fn products_update(
&self,
request: Attributes,
parent: &str,
name: &str
) -> AccountProductUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Inserts or updates the attributes of the product in a Manufacturer Center account.

Creates a product with the provided attributes. If the product already exists, then all attributes are replaced with the new ones. The checks at upload time are minimal. All required attributes need to be present for a product to be valid. Issues may show up later after the API has accepted a new upload for a product and it is possible to overwrite an existing valid product with an invalid product. To detect this, you should retrieve the product and check it for issues once the new version is available.

Uploaded attributes first need to be processed before they can be retrieved. Until then, new products will be unavailable, and retrieval of previously uploaded products will return the original state of the product.

Arguments

request - No description provided.

parent - Parent ID in the format accounts/{account_id}. account_id - The ID of the Manufacturer Center account.

name - Name in the format {target_country}:{content_language}:{product_id}. target_country - The target country of the product as a CLDR territory code (for example, US). content_language - The content language of the product as a two-letter ISO 639-1 language code (for example, en). product_id - The ID of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#id.

pub fn products_delete(
&self,
parent: &str,
name: &str
) -> AccountProductDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the product from a Manufacturer Center account.

Arguments

parent - Parent ID in the format accounts/{account_id}. account_id - The ID of the Manufacturer Center account.

name - Name in the format {target_country}:{content_language}:{product_id}. target_country - The target country of the product as a CLDR territory code (for example, US). content_language - The content language of the product as a two-letter ISO 639-1 language code (for example, en). product_id - The ID of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#id.

pub fn products_get(
&self,
parent: &str,
name: &str
) -> AccountProductGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the product from a Manufacturer Center account, including product issues.

A recently updated product takes around 15 minutes to process. Changes are only visible after it has been processed. While some issues may be available once the product has been processed, other issues may take days to appear.

Arguments

parent - Parent ID in the format accounts/{account_id}. account_id - The ID of the Manufacturer Center account.

name - Name in the format {target_country}:{content_language}:{product_id}. target_country - The target country of the product as a CLDR territory code (for example, US). content_language - The content language of the product as a two-letter ISO 639-1 language code (for example, en). product_id - The ID of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#id.