Struct google_manufacturers1::Attributes

pub struct Attributes {
    pub product_line: Option<String>,
    pub color: Option<String>,
    pub release_date: Option<String>,
    pub item_group_id: Option<String>,
    pub gtin: Option<Vec<String>>,
    pub excluded_destination: Option<Vec<String>>,
    pub flavor: Option<String>,
    pub target_client_id: Option<String>,
    pub scent: Option<String>,
    pub size: Option<String>,
    pub material: Option<String>,
    pub capacity: Option<Capacity>,
    pub title: Option<String>,
    pub pattern: Option<String>,
    pub disclosure_date: Option<String>,
    pub theme: Option<String>,
    pub product_detail: Option<Vec<ProductDetail>>,
    pub included_destination: Option<Vec<String>>,
    pub description: Option<String>,
    pub format: Option<String>,
    pub mpn: Option<String>,
    pub brand: Option<String>,
    pub suggested_retail_price: Option<Price>,
    pub product_name: Option<String>,
    pub feature_description: Option<Vec<FeatureDescription>>,
    pub size_system: Option<String>,
    pub size_type: Option<String>,
    pub additional_image_link: Option<Vec<Image>>,
    pub count: Option<Count>,
    pub gender: Option<String>,
    pub product_page_url: Option<String>,
    pub image_link: Option<Image>,
    pub product_type: Option<Vec<String>>,
    pub video_link: Option<Vec<String>>,
    pub age_group: Option<String>,
}

Attributes of the product. For more information, see https://support.google.com/manufacturers/answer/6124116.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

products update accounts (request)

Fields
product_line: Option<String>

The name of the group of products related to the product. For more information, see https://support.google.com/manufacturers/answer/6124116#productline.

color: Option<String>

The color of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#color.

release_date: Option<String>

The release date of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#release.

item_group_id: Option<String>

The item group id of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#itemgroupid.

gtin: Option<Vec<String>>

The Global Trade Item Number (GTIN) of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#gtin.

excluded_destination: Option<Vec<String>>

A list of excluded destinations.

flavor: Option<String>

The flavor of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#flavor.

target_client_id: Option<String>

The target client id. Should only be used in the accounts of the data partners.

scent: Option<String>

The scent of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#scent.

size: Option<String>

The size of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#size.

material: Option<String>

The material of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#material.

capacity: Option<Capacity>

The capacity of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#capacity.

title: Option<String>

The title of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#title.

pattern: Option<String>

The pattern of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#pattern.

disclosure_date: Option<String>

The disclosure date of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#disclosure.

theme: Option<String>

The theme of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#theme.

product_detail: Option<Vec<ProductDetail>>

The details of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#productdetail.

included_destination: Option<Vec<String>>

A list of included destinations.

description: Option<String>

The description of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#description.

format: Option<String>

The format of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#format.

mpn: Option<String>

The Manufacturer Part Number (MPN) of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#mpn.

brand: Option<String>

The brand name of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#brand.

suggested_retail_price: Option<Price>

The suggested retail price (MSRP) of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#price.

product_name: Option<String>

The canonical name of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#productname.

feature_description: Option<Vec<FeatureDescription>>

The rich format description of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#featuredesc.

size_system: Option<String>

The size system of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#sizesystem.

size_type: Option<String>

The size type of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#sizetype.

additional_image_link: Option<Vec<Image>>

The additional images of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#addlimage.

count: Option<Count>

The count of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#count.

gender: Option<String>

The target gender of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#gender.

product_page_url: Option<String>

The URL of the detail page of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#productpage.

image_link: Option<Image>

The image of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#image.

product_type: Option<Vec<String>>

The type or category of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#producttype.

video_link: Option<Vec<String>>

The videos of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#video.

age_group: Option<String>

The target age group of the product. For more information, see https://support.google.com/manufacturers/answer/6124116#agegroup.